Optimize merge function in graphql-anywhere.
This is extremely hot code, so small improvements matter.

Key observation: dest tends to be the current result object, which can
have a lot of keys that are irrelevant to the keys from src that we're
trying to merge, so the real win here is to avoid computing and iterating
over the keys of dest.